requirements.txt
PyGithub
Pillow